refactor(cli): migrate inference-config.js to TypeScript (#1265)
## Summary
- Convert `bin/lib/inference-config.js` (143 lines) to
`src/lib/inference-config.ts`
- Typed interfaces: `ProviderSelectionConfig`, `GatewayInference`
- All 3 exports are pure — straightforward full conversion
- Co-locate tests: `test/inference-config.test.js` →
`src/lib/inference-config.test.ts`

Stacked on #1240. 614 CLI tests pass. Coverage ratchet passes.

Relates to #924 (shell consolidation).
